requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
shap>=0.43
matplotlib>=3.7
seaborn>=0.13
plotly>=5.15
streamlit>=1.28
fastapi>=0.104
uvicorn>=0.24
joblib>=1.3
pydantic>=2.0
python-telegram-bot>=21.0